+++ react/ProductSummary.js
@@ -79,12 +79,12 @@ class ProductSummary extends Component {
                   />
                 </DiscountBadge>
               ) : (
-                  <img
-                    className="vtex-product-summary__image"
-                    alt={product.productName}
-                    src={product.sku.image.imageUrl}
-                  />
-                )}
+                <img
+                  className="vtex-product-summary__image"
+                  alt={product.productName}
+                  src={product.sku.image.imageUrl}
+                />
+              )}
             </div>
             <div className="vtex-product-summary__name-container flex items-center justify-center near-black">
               <ProductName
@@ -107,12 +107,12 @@ class ProductSummary extends Component {
           <div className="vtex-product-summary__buy-button-container pv2">
             {!hideBuyButton &&
               (!showButtonOnHover || this.state.isHovering) && (
-                <div className="vtex-product-summary__buy-button center">
-                  <BuyButton skuId={product.sku.itemId} isOneClickBuy={isOneClickBuy}>
-                    {buyButtonText || <FormattedMessage id="button-label" />}
-                  </BuyButton>
-                </div>
-              )}
+              <div className="vtex-product-summary__buy-button center">
+                <BuyButton skuId={product.sku.itemId} isOneClickBuy={isOneClickBuy}>
+                  {buyButtonText || <FormattedMessage id="button-label" />}
+                </BuyButton>
+              </div>
+            )}
           </div>
         </div>
       </div>
@@ -121,52 +121,38 @@ class ProductSummary extends Component {
 }
 
 const defaultSchema = {
-  title: 'Product Summary',
-  description: 'The product summary showing the main product informations',
+  title: 'editor.productSummary.title',
+  description: 'editor.productSummary.description',
   type: 'object',
   properties: {
-    showListPrice: {
-      type: 'boolean',
-      title: "Show product's list price",
-      default: true,
-    },
     isOneClickBuy: {
       type: 'boolean',
-      title: "Should redirect to checkout after clicking on buy",
+      title: 'editor.productSummary.isOneClickBuy.title',
       default: false,
-    },
-    showLabels: {
-      type: 'boolean',
-      title: "Show product's prices' labels",
-      default: true,
-    },
-    showInstallments: {
-      type: 'boolean',
-      title: "Show product's payment installments",
-      default: true,
     },
     showBadge: {
       type: 'boolean',
-      title: 'Show the discount badge',
+      title: 'editor.productSummary.showBadge.title',
       default: true,
     },
     badgeText: {
       type: 'string',
-      title: "Badge's text",
+      title: 'editor.productSummary.badgeText.title',
     },
     buyButtonText: {
       type: 'string',
-      title: "Custom buy button's text",
+      title: 'editor.productSummary.buyButtonText.title',
     },
     hideBuyButton: {
       type: 'boolean',
-      title: 'Hides the buy button completely',
+      title: 'editor.productSummary.hideBuyButton.title',
       default: false,
     },
     showButtonOnHover: {
       type: 'boolean',
-      title: 'Show the buy button only on hover',
+      title: 'editor.productSummary.showButtonOnHover.title',
     },
+    ...ProductPrice.schema.properties,
   },
 }