+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Added isOneClickBuy attribute.
-- Add internationalization into product summary schema
+- Add internationalization into `ProductSummary` schema
 
 ### Changed
 - `ProductSummary` schema to reuse the `ProductPrice` one